# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sourdough Bread Ordering Business Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  priceCents: number;
  active: boolean;
}

export interface CreateProductBody {
  name: string;
  /** @nullable */
  description?: string | null;
  priceCents: number;
  active?: boolean;
}

export interface UpdateProductBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  priceCents?: number;
  active?: boolean;
}

export interface BakingDay {
  id: number;
  date: string;
  productId: number;
  totalAvailable: number;
  reservedCount: number;
}

export interface BakingDayWithStats {
  id: number;
  date: string;
  productId: number;
  totalAvailable: number;
  reservedCount: number;
  paidCount: number;
  pendingCount: number;
  remaining: number;
  productName: string;
}

export interface CreateBakingDayBody {
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  productId: number;
  totalAvailable: number;
}

export interface UpdateBakingDayBody {
  totalAvailable?: number;
  date?: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending_payment: "pending_payment",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  whatsappNumber: string;
  /** @nullable */
  customerName?: string | null;
  bakingDayId: number;
  quantity: number;
  status: OrderStatus;
  /** @nullable */
  yocoPaymentId?: string | null;
  /** @nullable */
  yocoCheckoutId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type OrderWithDetailsStatus =
  (typeof OrderWithDetailsStatus)[keyof typeof OrderWithDetailsStatus];

export const OrderWithDetailsStatus = {
  pending_payment: "pending_payment",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export interface OrderWithDetails {
  id: number;
  whatsappNumber: string;
  /** @nullable */
  customerName?: string | null;
  bakingDayId: number;
  bakingDayDate: string;
  quantity: number;
  status: OrderWithDetailsStatus;
  /** @nullable */
  yocoPaymentId?: string | null;
  /** @nullable */
  yocoCheckoutId?: string | null;
  totalAmountCents: number;
  productName: string;
  createdAt: string;
  updatedAt: string;
}

export interface BakerSummary {
  totalOrdersToday: number;
  paidToday: number;
  pendingToday: number;
  upcomingDays: BakingDayWithStats[];
  recentOrders: OrderWithDetails[];
}

export type ListBakingDaysParams = {
  upcoming?: boolean;
};

export type ListOrdersParams = {
  bakingDayId?: number;
  status?: ListOrdersStatus;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending_payment: "pending_payment",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export type WhatsappWebhookBody = {
  From?: string;
  Body?: string;
};

export type YocoWebhookBody = { [key: string]: unknown };