# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sourdough Bread Ordering Business Manager API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns overview stats for the baker dashboard
 * @summary Get dashboard summary
 */
export const GetBakerSummaryResponse = zod.object({
  totalOrdersToday: zod.number(),
  paidToday: zod.number(),
  pendingToday: zod.number(),
  upcomingDays: zod.array(
    zod.object({
      id: zod.number(),
      date: zod.string(),
      productId: zod.number(),
      totalAvailable: zod.number(),
      reservedCount: zod.number(),
      paidCount: zod.number(),
      pendingCount: zod.number(),
      remaining: zod.number(),
      productName: zod.string(),
    }),
  ),
  recentOrders: zod.array(
    zod.object({
      id: zod.number(),
      whatsappNumber: zod.string(),
      customerName: zod.string().nullish(),
      bakingDayId: zod.number(),
      bakingDayDate: zod.string(),
      quantity: zod.number(),
      status: zod.enum(["pending_payment", "paid", "cancelled"]),
      yocoPaymentId: zod.string().nullish(),
      yocoCheckoutId: zod.string().nullish(),
      totalAmountCents: zod.number(),
      productName: zod.string(),
      createdAt: zod.string(),
      updatedAt: zod.string(),
    }),
  ),
});

/**
 * @summary List all products
 */
export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  priceCents: zod.number(),
  active: zod.boolean(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Create a new product
 */
export const CreateProductBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
  priceCents: zod.number(),
  active: zod.boolean().optional(),
});

/**
 * @summary Update a product
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProductBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().nullish(),
  priceCents: zod.number().optional(),
  active: zod.boolean().optional(),
});

export const UpdateProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  priceCents: zod.number(),
  active: zod.boolean(),
});

/**
 * @summary List all baking days (upcoming and past)
 */
export const ListBakingDaysQueryParams = zod.object({
  upcoming: zod.coerce.boolean().optional(),
});

export const ListBakingDaysResponseItem = zod.object({
  id: zod.number(),
  date: zod.string(),
  productId: zod.number(),
  totalAvailable: zod.number(),
  reservedCount: zod.number(),
  paidCount: zod.number(),
  pendingCount: zod.number(),
  remaining: zod.number(),
  productName: zod.string(),
});
export const ListBakingDaysResponse = zod.array(ListBakingDaysResponseItem);

/**
 * @summary Create a new baking day
 */
export const CreateBakingDayBody = zod.object({
  date: zod.string().describe("ISO date string (YYYY-MM-DD)"),
  productId: zod.number(),
  totalAvailable: zod.number(),
});

/**
 * @summary Update a baking day's availability
 */
export const UpdateBakingDayParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBakingDayBody = zod.object({
  totalAvailable: zod.number().optional(),
  date: zod.string().optional(),
});

export const UpdateBakingDayResponse = zod.object({
  id: zod.number(),
  date: zod.string(),
  productId: zod.number(),
  totalAvailable: zod.number(),
  reservedCount: zod.number(),
});

/**
 * @summary Delete a baking day
 */
export const DeleteBakingDayParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all orders with optional filters
 */
export const ListOrdersQueryParams = zod.object({
  bakingDayId: zod.coerce.number().optional(),
  status: zod.enum(["pending_payment", "paid", "cancelled"]).optional(),
});

export const ListOrdersResponseItem = zod.object({
  id: zod.number(),
  whatsappNumber: zod.string(),
  customerName: zod.string().nullish(),
  bakingDayId: zod.number(),
  bakingDayDate: zod.string(),
  quantity: zod.number(),
  status: zod.enum(["pending_payment", "paid", "cancelled"]),
  yocoPaymentId: zod.string().nullish(),
  yocoCheckoutId: zod.string().nullish(),
  totalAmountCents: zod.number(),
  productName: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Manually cancel an order
 */
export const CancelOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const CancelOrderResponse = zod.object({
  id: zod.number(),
  whatsappNumber: zod.string(),
  customerName: zod.string().nullish(),
  bakingDayId: zod.number(),
  quantity: zod.number(),
  status: zod.enum(["pending_payment", "paid", "cancelled"]),
  yocoPaymentId: zod.string().nullish(),
  yocoCheckoutId: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Yoco payment webhook
 */
export const YocoWebhookBody = zod.object({}).passthrough();