# lib/api-zod/src/generated/types/bakerSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sourdough Bread Ordering Business Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { BakingDayWithStats } from "./bakingDayWithStats";
import type { OrderWithDetails } from "./orderWithDetails";

export interface BakerSummary {
  totalOrdersToday: number;
  paidToday: number;
  pendingToday: number;
  upcomingDays: BakingDayWithStats[];
  recentOrders: OrderWithDetails[];
}