# lib/api-zod/src/generated/types/bakingDay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sourdough Bread Ordering Business Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface BakingDay {
  id: number;
  date: string;
  productId: number;
  totalAvailable: number;
  reservedCount: number;
}