# lib/api-zod/src/generated/types/bakingDayWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sourdough Bread Ordering Business Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface BakingDayWithStats {
  id: number;
  date: string;
  productId: number;
  totalAvailable: number;
  reservedCount: number;
  paidCount: number;
  pendingCount: number;
  remaining: number;
  productName: string;
}