# lib/api-zod/src/generated/types/createBakingDayBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sourdough Bread Ordering Business Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBakingDayBody {
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  productId: number;
  totalAvailable: number;
}