# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sourdough Bread Ordering Business Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  whatsappNumber: string;
  /** @nullable */
  customerName?: string | null;
  bakingDayId: number;
  quantity: number;
  status: OrderStatus;
  /** @nullable */
  yocoPaymentId?: string | null;
  /** @nullable */
  yocoCheckoutId?: string | null;
  createdAt: string;
  updatedAt: string;
}